🪟 🚦 Cleanup connection e2e test page objects (#23205)
* Group streams table tests into different groups

* streamTable.spec -> createConnection.spec.ts

* Group connection page objects in pages/connection and rename

* Clean up auto-detect schema tests to use page objects as * imports

* Fix createConnectionPageObject name

* Simplify function names in page objects

* Split streams replication into connetion form and streams table page objects

* Cleanup connection spec

* Move abstract StreamsTable class to interface and base class
