+++ airbyte-webapp-e2e-tests/cypress/pages/connection/connectionPageObject.ts
@@ -1,12 +1,17 @@
 import { Connection } from "commands/api/types";
 import { getWorkspaceId } from "commands/api/workspace";
 
+const replicationTab = "div[data-id='replication-step']";
 const syncEnabledSwitch = "[data-testid='enabledControl-switch']";
 
-export const visitConnectionPage = (connection: Connection, tab = "") => {
+export const visit = (connection: Connection, tab = "") => {
   cy.intercept("**/web_backend/connections/get").as("getConnection");
   cy.visit(`/workspaces/${getWorkspaceId()}/connections/${connection.connectionId}/${tab}`);
   cy.wait("@getConnection", { timeout: 20000 });
+};
+
+export const goToReplicationTab = () => {
+  cy.get(replicationTab).click();
 };
 
 export const getSyncEnabledSwitch = () => cy.get(syncEnabledSwitch);
