+++ airbyte-webapp-e2e-tests/cypress/pages/connection/connectionSettingsPageObject.ts
