+++ airbyte-webapp-e2e-tests/cypress/pages/connection/createConnectionPageObject.ts
