+++ airbyte-webapp-e2e-tests/cypress/pages/connection/streamsTablePageObject/index.ts
@@ -1,0 +1,3 @@
+import { LegacyStreamsTablePageObject } from "./LegacyStreamsTablePageObject";
+
+export default new LegacyStreamsTablePageObject();
